whiteboard image addeed and each methods completed
