--- Data Structures/LinkedList/LinkedList/Program.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace LinkedList
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-        }
-    }
-}
