+++ Data-Structures/LinkedList/LinkedList.sln
@@ -1,0 +1,23 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedList", "LinkedList\LinkedList.csproj", "{55F562F6-E9A4-46BC-9DBF-87664CFA4E8F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeChallenge06Test", "CodeChallenge06Test\CodeChallenge06Test.csproj", "{B7084E42-21A0-4A7C-A4ED-F6938C1E7044}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{55F562F6-E9A4-46BC-9DBF-87664CFA4E8F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55F562F6-E9A4-46BC-9DBF-87664CFA4E8F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55F562F6-E9A4-46BC-9DBF-87664CFA4E8F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55F562F6-E9A4-46BC-9DBF-87664CFA4E8F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B7084E42-21A0-4A7C-A4ED-F6938C1E7044}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B7084E42-21A0-4A7C-A4ED-F6938C1E7044}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B7084E42-21A0-4A7C-A4ED-F6938C1E7044}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B7084E42-21A0-4A7C-A4ED-F6938C1E7044}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
