+++ Data-Structures/LinkedList/LinkedList/Classes/Node.cs
