+++ Data-Structures/LinkedList/LinkedList/Program.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace LinkedList
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+        }
+    }
+}
